Created spectrum bar display for swing application

+++ GriegSwing/src/pl/edu/agh/ki/grieg/swing/util/Utils.java
@@ -6,8 +6,8 @@ public class Utils {
         // non-instantiable
     }
     
-    public static float dB(float x) {
-        return 10 * (float) Math.log10(x);
+    public static double dB(double x) {
+        return 10 * Math.log10(x);
     }
     
     public static double clamp(double val, double min, double max) {
